+++ .luarc.json
@@ -1,0 +1,5 @@
+{
+    "diagnostics.globals": [
+        "vim"
+    ]
+}
