--- after/plugin/dap.lua
@@ -1,4 +1,0 @@
-vim.keymap.set("n", "<leader>db", "<cmd> DapToggleBreakpoint <CR>")
-vim.keymap.set("n", "<leader>dpr", function()
-	require("dap-python").test_method()
-end)
