--- after/plugin/fugitive.lua
@@ -1,2 +1,0 @@
-vim.keymap.set("n", "<leader>gs", vim.cmd.Git);
-
