--- after/plugin/harpoon.lua
@@ -1,51 +1,0 @@
-local harpoon = require("harpoon")
-
--- REQUIRED
-harpoon:setup({
-	settings = {
-		save_on_toggle = true,
-	},
-}) -- REQUIRED
-
-vim.keymap.set("n", "<leader>ha", function()
-	harpoon:list():append()
-end)
-vim.keymap.set("n", "<leader>hs", function()
-	harpoon.ui:toggle_quick_menu(harpoon:list())
-end)
-
-vim.keymap.set("n", "<leader>1", function()
-	harpoon:list():select(1)
-end)
-vim.keymap.set("n", "<leader>2", function()
-	harpoon:list():select(2)
-end)
-vim.keymap.set("n", "<leader>3", function()
-	harpoon:list():select(3)
-end)
-vim.keymap.set("n", "<leader>4", function()
-	harpoon:list():select(4)
-end)
-
--- Toggle previous & next buffers stored within Harpoon list
-vim.keymap.set("n", "<C-S-P>", function()
-	harpoon:list():prev()
-end)
-vim.keymap.set("n", "<C-S-N>", function()
-	harpoon:list():next()
-end)
-
-harpoon:extend({
-	UI_CREATE = function(cx)
-		vim.keymap.set("n", "<C-v>", function()
-			harpoon.ui:select_menu_item({ vsplit = true })
-		end, { desc = "Open as vertical split", buffer = cx.bufnr })
-		vim.keymap.set("n", "<C-x>", function()
-			harpoon.ui:select_menu_item({ split = true })
-		end, { desc = "Open as split", buffer = cx.bufnr })
-		vim.keymap.set("n", "<C-t>", function()
-			harpoon.ui:select_menu_item({ tabedit = true })
-		end, { desc = "Open in new tab", buffer = cx.bufnr })
-	end,
-	-- Also tried using NAVIGATE event here instead of extensions.builtins, same luck
-})
