--- after/plugin/indent_blank_lines.lua
@@ -1,7 +1,0 @@
-require("ibl").setup({
-	scope = {
-		enabled = false,
-		show_start = true,
-		show_end = true,
-	},
-})
