+++ init.lua
@@ -1,3 +1,1 @@
-require("theprimeagen")
 require("vsubota")
-
