--- lua/theprimeagen/init.lua
@@ -1,3 +1,0 @@
-require("theprimeagen.remap")
-require("theprimeagen.set")
-
