--- lua/vsubota/mouse.lua
@@ -1,3 +1,0 @@
-
-vim.opt.mouse = ""
-
