+++ lua/vsubota/plugins/commentary.lua
@@ -1,0 +1,1 @@
+return {"tpope/vim-commentary"}
