+++ lua/vsubota/plugins/dap-go.lua
@@ -1,0 +1,9 @@
+return {
+    "leoluz/nvim-dap-go",
+    ft = { "go" },
+    event = "VeryLazy",
+    requires = "mfussenegger/nvim-dap",
+    config = function()
+        require("dap-go").setup()
+    end,
+}
