+++ lua/vsubota/plugins/dap-ui.lua
@@ -1,0 +1,21 @@
+return {
+    "rcarriga/nvim-dap-ui",
+    dependencies = "mfussenegger/nvim-dap",
+    event = "VeryLazy",
+    library = { plugins = { "nvim-dap-ui" }, types = true },
+    config = function()
+        local dap, dapui = require("dap"), require("dapui")
+        dapui.setup()
+
+        dap.listeners.before.attach.dapui_config = function()
+            dapui.open()
+        end
+        dap.listeners.before.launch.dapui_config = function()
+            dapui.open()
+        end
+        dap.listeners.before.event_terminated.dapui_config = function()
+            -- dapui.close()
+        end
+        dap.listeners.before.event_exited.dapui_config = function() end
+    end,
+}
