+++ lua/vsubota/plugins/dap.lua
@@ -1,0 +1,7 @@
+return {
+    "mfussenegger/nvim-dap",
+    event = "VeryLazy",
+    config = function() 
+        vim.keymap.set("n", "<leader>db", "<cmd> DapToggleBreakpoint <CR>")
+    end
+}
