+++ lua/vsubota/plugins/diffview.lua
@@ -1,0 +1,1 @@
+return {"sindrets/diffview.nvim"}
