+++ lua/vsubota/plugins/dressing.lua
@@ -1,0 +1,1 @@
+return {"stevearc/dressing.nvim"}
