+++ lua/vsubota/plugins/lazy-git.lua
@@ -1,0 +1,11 @@
+return {
+    "kdheepak/lazygit.nvim",
+    event = "VeryLazy",
+    -- optional for floating window border decoration
+    dependencies = {
+        "nvim-lua/plenary.nvim",
+    },
+    config = function()
+        vim.keymap.set("n", "<leader>lg", "<cmd>LazyGit<CR>")
+    end
+}
