+++ lua/vsubota/plugins/lualine.lua
@@ -1,0 +1,7 @@
+return {
+    'nvim-lualine/lualine.nvim',
+    dependencies = { 'nvim-tree/nvim-web-devicons' },
+    config = function()
+        require('lualine').setup()
+    end
+}
