+++ lua/vsubota/plugins/neodev.lua
@@ -1,0 +1,1 @@
+return { "folke/neodev.nvim", opts = {} }
