+++ lua/vsubota/plugins/treesitter.lua
@@ -1,0 +1,61 @@
+local M = {
+    "nvim-treesitter/nvim-treesitter",
+    event = "VeryLazy",
+    build = function()
+        require("nvim-treesitter.install").update({ with_sync = true })()
+    end,
+    config = function()
+        require("nvim-treesitter.configs").setup({
+            -- A list of parser names, or "all" (the five listed parsers should always be installed)
+            ensure_installed = {
+                "help",
+                "lua",
+                "vim",
+                "vimdoc",
+                "go",
+                "gomod",
+                "gosum",
+                "python",
+                "yaml",
+                "json",
+                "markdown_inline",
+                "markdown",
+            },
+
+            -- enable indentation
+            indent = { enable = true },
+
+            -- Install parsers synchronously (only applied to `ensure_installed`)
+            sync_install = false,
+
+            -- Automatically install missing parsers when entering buffer
+            -- Recommendation: set to false if you don't have `tree-sitter` CLI installed locally
+            auto_install = true,
+
+            ignore_install = { "help" },
+
+            highlight = {
+                enable = true,
+
+                -- Setting this to true will run `:h syntax` and tree-sitter at the same time.
+                -- Set this to `true` if you depend on 'syntax' being enabled (like for indentation).
+                -- Using this option may slow down your editor, and you may see some duplicate highlights.
+                -- Instead of true it can also be a list of languages
+                additional_vim_regex_highlighting = false,
+            },
+
+            incremental_selection = {
+                enable = true,
+                keymaps = {
+                    init_selection = "<C-space>",
+                    node_incremental = "<C-space>",
+                    scope_incremental = false,
+                    node_decremental = "<bs>",
+                },
+            },
+        })
+
+    end
+}
+
+return { M }
