+++ lua/vsubota/plugins/web-devicons.lua
@@ -1,0 +1,1 @@
+return {"nvim-tree/nvim-web-devicons"}
